Added bootlogo based on Wikipedia Tux.svg image
This is a derived work from http://commons.wikimedia.org/wiki/File:Tux.svg
The original Copyright states the following:

The copyright holder of this file allows anyone to use it for any purpose,
provided that the copyright holder is properly attributed. Redistribution,
derivative work, commercial use, and all other use is permitted.

The same applies to this derived work.



Signed-off-by: Christian Ege <k4230r6@gmail.com>