This display a continues bootlogo in 720p during boot
In the previous version the bootlogo disappeared and the monitor went black.
This was caused by switching from 720p to 1080p. If we only set the the
resolution to 720p in the boot.ini the psplash gets confused and corrupted
during boot. So I added this aaa.sh initscript which calls fbset before
calling psplash this will allows psplash to use the correct resolution.


Signed-off-by: Christian Ege <k4230r6@gmail.com>