Moved amlsetfb to a separate package
The previous approach of appending the initscripts package caused troubles
when using the amlogic layer in combination with another layer. In this case
any other machine could not satisfy the dependency to amlsetfb. What caused
the build to fail.

Due to the fact that amlsetfb depends on the framebuffer helper tools from
meta-openembedded we need some extra checks.

Signed-off-by: Christian Ege <k4230r6@gmail.com>